# swagger/swagger.json
{
  "swagger": "2.0" ,
  "info": {
    "version": "1.0.0" ,
    "title": "Micro Credits Backend" ,
    "description": "Micro Credits Backend" ,
    "license": {
      "name": "MIT" ,
      "url": "https://opensource.org/licenses/MIT"
    }
  } ,
  "host": "ec2-54-171-157-12.eu-west-1.compute.amazonaws.com:3005",
...
  "schemes": [
    "http"
  ] ,
  "consumes": [
    "application/json"
  ] ,
  "produces": [
    "application/json"
  ] ,
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security":[
    {
      "api_key":[]
    }
  ],
  "paths":{
    "/credits": {
      "get":{
        "tags": [
          "Credits"
        ],
        "summary": "Gets all various types of credits available",
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Credits"
        ],
        "parameters": [
          {
            "name": "credits",
            "in": "body",
            "description": "Credit data",
            "schema": {
              "$ref": "#/definitions/creditData"
            }
          }
        ],
        "summary": "Creates new credit record",
        "responses": {
          "200": {
            "description": "New credit record is created" ,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/credits/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of credit record that we want",
          "type": "string"
        }
      ],
      "put": {
        "summary": "Update credit record with give ID",
        "tags": [
          "Credits"
        ],
        "parameters": [
          {
            "name": "credit",
            "in": "body",
            "description": "Credit record with new values of properties",
            "schema": {
              "$ref": "#/definitions/creditUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credit record updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/country": {
      "get": {
        "tags": [
          "Country"
        ] ,
        "summary": "Get all countries" ,
        "responses": {
          "200": {
            "description": "OK" ,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "country",
            "in": "body",
            "description": "Body of country that we want to create",
            "schema": {
              "$ref": "#/definitions/countryData"
            }
          }
        ],
        "summary": "Creates new Country",
        "responses": {
          "200": {
            "description": "New country is created" ,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/country/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of country that we want to find",
          "type": "string"
        }
      ],
      "delete": {
        "summary": "Delete country with given ID",
        "tags": [
          "Country"
        ],
        "responses": {
          "200": {
            "description": "Country is deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update country with give ID",
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "name": "country",
            "in": "body",
            "description": "Country with new values of properties",
            "schema": {
              "$ref": "#/definitions/countryUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Country is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Country"
        ],
        "summary": "Get country by id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
    "/state/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of state that we want to find",
          "type": "string"
        }
      ],
      "delete": {
        "summary": "Delete state with given ID",
        "tags": [
          "State"
        ],
        "responses": {
          "200": {
            "description": "State is deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update state with give ID",
        "tags": [
          "State"
        ],
        "parameters": [
          {
            "name": "state",
            "in": "body",
            "description": "State with new values of properties",
            "schema": {
              "$ref": "#/definitions/stateUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "State is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "State"
        ],
        "summary": "Get state by id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/state/country/{countryID}": {
      "parameters": [
        {
          "name": "countryID",
          "in": "path",
          "required": true,
          "description": "ID of country that we want to find states of",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "State"
        ],
        "summary": "Get states by country id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/city": {
      "get": {
        "tags": [
          "City"
        ] ,
        "summary": "Get all cities" ,
        "responses": {
          "200": {
            "description": "OK" ,
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      },
      "post": {
        "tags": [
          "City"
        ],
        "parameters": [
          {
            "name": "City",
            "in": "body",
            "description": "Body of city that we want to create",
            "schema": {
              "$ref": "#/definitions/cityData"
            }
          }
        ],
        "summary": "Creates new City",
        "responses": {
          "200": {
            "description": "New city is created" ,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/city/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of city that we want to find",
          "type": "string"
        }
      ],
      "delete": {
        "summary": "Delete city with given ID",
        "tags": [
          "City"
        ],
        "responses": {
          "200": {
            "description": "City is deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update city with give ID",
        "tags": [
          "City"
        ],
        "parameters": [
          {
            "name": "City",
            "in": "body",
            "description": "City with new values of properties",
            "schema": {
              "$ref": "#/definitions/cityUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "City is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "City"
        ],
        "summary": "Get city by id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/city/state/{stateID}": {
      "parameters": [
        {
          "name": "stateID",
          "in": "path",
          "required": true,
          "description": "ID of state that we want to find cities of",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "City"
        ],
        "summary": "Get cities by state id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/city/country/{countryID}": {
      "parameters": [
        {
          "name": "countryID",
          "in": "path",
          "required": true,
          "description": "ID of country that we want to find cities of",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "City"
        ],
        "summary": "Get cities by country id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "creditData": {
      "required": [
        "key",
        "value"
      ],
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "creditUpdate": {
      "required": [
        "value"
      ],
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
...
    "countryData": {
      "required": [
        "countryName"
      ],
      "properties": {
        "countryName": {
          "type": "string"
        }
      }
    },
    "countryUpdate": {
      "required": [
        "countryName"
      ],
      "properties": {
        "countryName": {
          "type": "string"
        }
      }
    },
...
    "stateUpdate": {
      "required": [
        "stateName"
      ],
      "properties": {
        "stateName": {
          "type": "string"
        }
      }
    },
    "cityData": {
      "required": [
        "cityName",
        "stateId"
      ],
      "properties": {
        "cityName": {
          "type": "string"
        },
        "stateId": {
          "type": "integer"
        }
      }
    },
    "cityUpdate": {
      "required": [
        "cityName"
      ],
      "properties": {
        "cityName": {
          "type": "string"
        }
      }
    },
...
  }
}